+++ src/utils/remark.ts
@@ -1,19 +1,9 @@
 import { h as _h, type Properties } from "hastscript";
-import type { Node, Paragraph as P } from "mdast";
-import type { Directives } from "mdast-util-directive";
-
-/** Checks if a node is a directive. */
-export function isNodeDirective(node: Node): node is Directives {
-	return (
-		node.type === "containerDirective" ||
-		node.type === "leafDirective" ||
-		node.type === "textDirective"
-	);
-}
+import type { Paragraph } from "mdast";
 
 /** From Astro Starlight: Function that generates an mdast HTML tree ready for conversion to HTML by rehype. */
 // biome-ignore lint/suspicious/noExplicitAny: allow any children
-export function h(el: string, attrs: Properties = {}, children: any[] = []): P {
+export function h(el: string, attrs: Properties = {}, children: any[] = []): Paragraph {
 	const { properties, tagName } = _h(el, attrs);
 	return {
 		children,
